Equally broken but hopefully internally better build
