+++ 1.3/UI/StaticPoisonDatabase.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DI_Harmacy
+{
+    public static class StaticPoisonDatabase
+    {
+        public static PoisonPolicyDatabase poisonPolicyDatabase;
+    }
+}
