+++ Diamond Industries Harmacy.csproj
@@ -118,16 +118,14 @@
     <Compile Include="1.3\Patches\HarmonyPatching.cs" />
     <Compile Include="1.3\Patches\RangedHarmonyPatch.cs" />
     <Compile Include="1.3\UI\PoisonPolicyDataBase.cs" />
-    <None Include="1.3\UI\OldDrugModification\PoisonPolicyDef.cs" />
+    <Compile Include="1.3\UI\PoisonPolicyDef.cs" />
     <Compile Include="1.3\UI\Dialog_ManagePoisonPolicies.cs" />
     <Compile Include="1.3\UI\PoisonPolicyEntry.cs" />
-    <None Include="1.3\UI\OldDrugModification\PawnColumnWorker_PoisonPolicy.cs" />
-    <None Include="1.3\UI\OldDrugModification\PoisonPolicy.cs" />
-    <None Include="1.3\UI\OldDrugModification\Pawn_PoisonPolicyTracker.cs" />
-    <None Include="1.3\UI\OldDrugModification\PoisonPolicyUIUtility.cs" />
+    <Compile Include="1.3\UI\PoisonPolicy.cs" />
+    <Compile Include="1.3\UI\PoisonPolicyUIUtility.cs" />
     <Compile Include="1.3\UI\Pawn_PoisonPolicyTracker.cs" />
     <Compile Include="1.3\UI\PawnColumnWorker_PoisonPolicy.cs" />
-    <Compile Include="1.3\UI\PoisonPolicy.cs" />
+    <Compile Include="1.3\UI\StaticPoisonDatabase.cs" />
     <Compile Include="1.3\Verbs\VerbPoison_Jump.cs" />
     <Compile Include="1.3\Verbs\Verbs.cs" />
     <None Include="DIH\ModSettings.cs" />
